Move control lines code to new files
